+++ lakefs_provider/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.47.0'
+__version__ = '0.48.0'
 
 ## This is needed to allow Airflow to pick up specific metadata fields it needs for certain features. We recognize it's a bit unclean to define these in multiple places, but at this point it's the only workaround if you'd like your custom conn type to show up in the Airflow UI.
 def get_provider_info():