+++ lakefs_provider/hooks/lakefs_hook.py
@@ -2,26 +2,14 @@ from typing import Any, Dict, IO, Iterator
 
 from lakefs_provider import __version__
 
-import lakefs_client
-from lakefs_client import models
-from lakefs_client.client import LakeFSClient
-from lakefs_client.model.object_stats import ObjectStats
-from lakefs_client.models import Merge
+import lakefs_sdk
+from lakefs_sdk import models
+from lakefs_sdk.client import LakeFSClient
+from lakefs_sdk.models.object_stats import ObjectStats
+from lakefs_sdk.models import Merge
 
 from airflow.exceptions import AirflowException
 from airflow.hooks.base import BaseHook
-
-
-def _commitAsDict(details):
-    return {
-        "id": details.id,
-        "parents": details.parents,
-        "committer": details.committer,
-        "message": details.message,
-        "creation_date": details.creation_date,
-        "meta_range_id": details.meta_range_id,
-        "metadata": details.metadata,
-    }
 
 
 class LakeFSHook(BaseHook):
@@ -50,8 +38,9 @@ class LakeFSHook(BaseHook):
 
     def get_conn(self) -> LakeFSClient:
         conn = self.get_connection(self.lakefs_conn_id)
-        configuration = lakefs_client.Configuration()
-        if conn.conn_type == "http" and conn.extra_dejson.get("access_key_id") and conn.extra_dejson.get("secret_access_key"):
+        configuration = lakefs_sdk.Configuration()
+        if conn.conn_type == "http" and conn.extra_dejson.get("access_key_id") and conn.extra_dejson.get(
+                "secret_access_key"):
             configuration.username = conn.extra_dejson.get("access_key_id")
             configuration.password = conn.extra_dejson.get("secret_access_key")
         else:
@@ -82,7 +71,6 @@ class LakeFSHook(BaseHook):
         ref = client.branches_api.create_branch(
             repository=repository, branch_creation=models.BranchCreation(name=name,
                                                                          source=source_branch))
-
         return ref
 
     def commit(self, repo: str, branch: str, msg: str, metadata: Dict[str, Any] = None) -> str:
@@ -92,9 +80,9 @@ class LakeFSHook(BaseHook):
             branch=branch,
             commit_creation=models.CommitCreation(message=msg, metadata=metadata))
 
-        return commit.get("id")
+        return commit.id
 
-    def upload(self, repo: str, branch: str, path: str, content: IO) -> str:
+    def upload(self, repo: str, branch: str, path: str, content: bytes) -> str:
         client = self.get_conn()
         upload = client.objects_api.upload_object(
             repository=repo,
@@ -113,7 +101,7 @@ class LakeFSHook(BaseHook):
             destination_branch=destination_branch,
             merge=Merge(message=msg, metadata=metadata))
 
-        return merge_result.get("reference")
+        return merge_result.reference
 
     def get_branch_commit_id(self, repo: str, name: str) -> str:
         client = self.get_conn()
@@ -122,54 +110,53 @@ class LakeFSHook(BaseHook):
 
     def get_commit(self, repo: str, ref: str) -> Dict[str, str]:
         client = self.get_conn()
-        # Actually ask for a log of length 1, because of treeverse/lakeFS#3436.
-        response = client.refs_api.log_commits(repo, ref, amount=1)
-        details = response.results[0]
-        return _commitAsDict(details)
+        commit = client.commits_api.get_commit(repo, ref)
+        return commit.to_dict()
 
-    def log_commits(self, repo: str, ref: str, size: int=100) -> Iterator:
+    def log_commits(self, repo: str, ref: str, size: int = 100) -> Iterator:
         """Yield commits of repo backwards from ref.
-
         Fetch size commits at a time."""
         client = self.get_conn()
-        log_commits = client.refs_api.log_commits
         after = ''
         while True:
-            response = log_commits(repo, ref, amount=size, after=after)
+            response = client.refs_api.log_commits(repo, ref, amount=size, after=after)
             for details in response.results:
-                yield _commitAsDict(details)
+                yield details.to_dict()
             if response.pagination is None or not response.pagination.has_more:
                 return
             after = response.pagination.next_offset
 
     def stat_object(self, repo: str, ref: str, path: str) -> ObjectStats:
         client = self.get_conn()
-        return client.objects_api.stat_object(repository=repo, ref=ref, path=path)
+        response = client.objects_api.stat_object(repository=repo, ref=ref, path=path)
+        return response.to_dict()
 
     def get_object(self, repo: str, ref: str, path: str) -> IO:
         client = self.get_conn()
         return client.objects_api.get_object(repository=repo, ref=ref, path=path)
 
-    def create_symlink_file(self, repo: str, branch: str, location: str = None)  -> str:
+    def create_symlink_file(self, repo: str, branch: str, location: str = None) -> str:
         client = self.get_conn()
 
         kwargs = {}
         if location:
             kwargs["location"] = location
 
-        return client.internal_api.create_symlink_file(repository=repo, branch=branch, **kwargs)["location"]
+        response = client.internal_api.create_symlink_file(repository=repo, branch=branch, **kwargs)
+        return response.location
 
     def delete_branch(self, repo: str, branch: str) -> str:
         client = self.get_conn()
         return client.branches_api.delete_branch(repository=repo, branch=branch)
 
     def test_connection(self):
-        """Test  Connection"""
+        """Test Connection"""
         conn = self.get_connection(self.lakefs_conn_id)
         import requests
         import json
-        url = conn.host+"/api/v1/auth/login"
-        if conn.conn_type == "http" and conn.extra_dejson.get("access_key_id") and conn.extra_dejson.get("secret_access_key"):
+        url = conn.host + "/api/v1/auth/login"
+        if conn.conn_type == "http" and conn.extra_dejson.get("access_key_id") and conn.extra_dejson.get(
+                "secret_access_key"):
             login = conn.extra_dejson.get("access_key_id")
             password = conn.extra_dejson.get("secret_access_key")
         else:
@@ -183,10 +170,10 @@ class LakeFSHook(BaseHook):
         response = requests.request("POST", url, headers=headers, data=payload)
         try:
             response.raise_for_status()
-            return True,"Connection Tested Successfully"
+            return True, "Connection Tested Successfully"
         except requests.exceptions.URLRequired as e:
-            return  False,str(e)
+            return False, str(e)
         except requests.exceptions.HTTPError as e:
-            return  False,str(e)
+            return False, str(e)
         except requests.exceptions.RequestException as e:
-            return  False,str(e)
+            return False, str(e)