+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG.md
 
+## 0.48.0
+
+  * Use new lakeFS Python SDK v0.113.0
+    Adopt calls and response from lakeFS to new API
+
 ## 0.46.2
 
   * Update lakeFS client SDK to v0.101.0.